Split renderer into text.rs and railroad.rs
